pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atlas-trading"
version = "0.1.0"
description = "Intraday backtesting + Alpaca paper trading scaffold for US stocks."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "alpaca-py>=0.25.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "python-dotenv>=1.0.0",
  "typer>=0.12.0",
]

[project.scripts]
atlas = "atlas.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
